## Database Layer    - Configure Hibernate JPA with EntityManagerFactory    - Add Flyway migrations for schema management    - Update all repositories to use JPA/Hibernate instead of in-memory storage    - Add persistence.xml configuration for JPA entities    - Implement BaseUserRepository with JPA operations
   ## Entity Updates
   - Add JPA annotations (@Entity, @Table, @Id, etc.) to all entities
   - Create UserHolder interface for User-based entities
   - Update relationships with @OneToMany, @ManyToOne, @ManyToMany
   - Configure cascade types and fetch strategies

   ## Service Layer Enhancements
   - Refactor TrainerServiceImpl.create() to properly set specialization from database
   - Add @Transactional annotations for transaction management
   - Update AbstractUserService with transaction rollback configuration
   - Fix service layer to work with JPA entities

   ## Security Implementation
   - Add SecurityGymFacade decorator with authentication and authorization
   - Create UserContext for managing logged-in user state
   - Implement SecurityService for authentication operations
   - Add custom exceptions: UnauthorizedException, ForbiddenException
   - Define UserRole enum (TRAINEE, TRAINER)
   - Create GymFacadeInterface to enable facade pattern

   ## Configuration
   - Update JpaConfig with EntityManagerFactory and TransactionManager beans
   - Configure Flyway for database migrations
   - Add application.properties for database connection settings
   - Remove obsolete Spring test configuration files

   ## Bug Fixes
   - Fix trainer specialization null issue by setting it during entity creation
   - Resolve mapping issues in TrainerMapper

+++ src/main/java/epam/gym/infrastructure/entities/Training.java
@@ -51,6 +51,6 @@ public class Training {
 
     @NotNull
     @Column(name = "training_duration", nullable = false)
-    Long trainingDuration;
+    Integer trainingDuration;
 
 }
